requirements.txt
alembic==1.8.1
anyio==3.6.1
Babel==2.10.3
beautifulsoup4==4.12.2
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.0
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Modals==0.5.0
Flask-Moment==1.0.4
Flask-Reuploaded==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
http-ece==1.1.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
onesignal-sdk==2.0.0
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
py-vapid==1.8.2
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.2.1
pywebpush==1.14.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.39
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
